+++ R/benchmark.R
@@ -20,6 +20,7 @@
 #'   Performance measures for all tasks.
 #'   If missing, the default measure of the first task is used.
 #' @template arg_keep_pred
+#' @template arg_keep_extract
 #' @template arg_models
 #' @template arg_showinfo
 #' @return [BenchmarkResult].
@@ -38,7 +39,8 @@
 #' plotBMRRanksAsBarChart(bmr, pos = "stack")
 #' friedmanTestBMR(bmr)
 #' friedmanPostHocTestBMR(bmr, p.value = 0.05)
-benchmark = function(learners, tasks, resamplings, measures, keep.pred = TRUE, models = TRUE, show.info = getMlrOption("show.info")) {
+benchmark = function(learners, tasks, resamplings, measures, keep.pred = TRUE,
+  keep.extract = FALSE, models = FALSE, show.info = getMlrOption("show.info")) {
 
   learners = ensureBenchmarkLearners(learners)
   tasks = ensureBenchmarkTasks(tasks)
@@ -56,7 +58,8 @@ benchmark = function(learners, tasks, resamplings, measures, keep.pred = TRUE, m
     task = grid$task,
     learner = grid$learner,
     more.args = list(learners = learners, tasks = tasks, resamplings = resamplings,
-      measures = measures, keep.pred = keep.pred, models = models, show.info = show.info),
+      measures = measures, keep.pred = keep.pred, models = models, show.info = show.info,
+      keep.extract = keep.extract),
     level = plevel
   )
   results.by.task = split(results, unlist(grid$task))
@@ -100,14 +103,20 @@ benchmark = function(learners, tasks, resamplings, measures, keep.pred = TRUE, m
 #' @family benchmark
 NULL
 
-
-benchmarkParallel = function(task, learner, learners, tasks, resamplings, measures, keep.pred = TRUE, models = TRUE, show.info) {
+benchmarkParallel = function(task, learner, learners, tasks, resamplings,
+  measures, keep.pred = TRUE, keep.extract = FALSE, models = FALSE, show.info) {
   setSlaveOptions()
   if (show.info) {
     messagef("Task: %s, Learner: %s", task, learner)
   }
   lrn = learners[[learner]]
-  extract.this = getExtractor(lrn)
+  if (isTRUE(keep.extract)) {
+    extract.this = getExtractor(lrn)
+  } else {
+    extract.this = function(model) {
+      NULL
+    }
+  }
   r = resample(lrn, tasks[[task]], resamplings[[task]],
     measures = measures, models = models, extract = extract.this, keep.pred = keep.pred, show.info = show.info)
   # store used learner in result