+++ man-roxygen/arg_keep_pred.R
@@ -2,6 +2,5 @@
 #'   Keep the prediction data in the `pred` slot of the result object.
 #'   If you do many experiments (on larger data sets) these objects might unnecessarily increase
 #'   object size / mem usage, if you do not really need them.
-#'   In this case you can set this argument to `FALSE`.
-#'   Default is `TRUE`.
+#'   The default is set to `TRUE`.
 #' @md