+++ man-roxygen/arg_models.R
@@ -1,4 +1,4 @@
 #' @param models (`logical(1)`)\cr
 #'   Should all fitted models be stored in the [ResampleResult]?
-#'   Default is `TRUE`.
+#'   Default is `FALSE`.
 #' @md